# [IMP] purchase_stock: picking return - perf