Before this commit:
When doing a return on a picking with multiple hundreds of products,
the return processing would timeout.

The main slowness culprit was
`line.move_ids.mapped('returned_move_ids')`
which was going through multiple one2many relations and prefetching a
lot of useless data.

Empirical testing shows the execution time of _compute_picking being decreased by a factor
of 30x on very large datasets.

opw-2161341