@Whenrow 2020-01-03
could you try this : 
``` python
data = self.env['stock.move'].read_group(
                [('purchase_line_id', 'in', order.order_line.ids)],
                ['id', 'picking_id'], ['id', 'picking_id'],
                lazy=False
            )
            ids = {}
            for move in data:
                ids.add(move['id'])

            data2 = self.env['stock.move'].read_group(
                [('origin_returned_move_id', 'in', ids)],
                ['picking_id'], ['picking_id'],
                lazy=False
            )

# next get all `picking_id` from data1 and data2
pickings = ....
```