@mao-odoo 2020-01-06
@Whenrow 
The read_group method does not allow for grouping on 'id' (ambiguous query), and of course, without the possibility to group_by on it, the whole point of using the read_group is moot here.