@sle-odoo 2020-01-06
@mao-odoo are you sure the issue stems from the use of mapped and the prefetching ? are both mapped used in this methods slow? if it's indeed the case and due to the prefetch, did you try disabling the prefetch at this call?

also, i tried to adapt the method last week but i was not able to get it working, if the mapped are indeed slow maybe you can try to search the records out of the loop. here's the snipped i was working on:
``` python
moves_per_line = defaultdict(lambda: self.env['stock.move'])
returned_moves_per_move = defaultdict(lambda: self.env['stock.move'])

line_ids = orders.mapped('order_line').ids
moves = self.env['stock.move'].search([('purchase_line_id', 'in', line_ids)])
for move in moves:
    moves_per_line[move.purchase_line_id] |= move
returned_moves = self.env['stock.move'].search([('origin_returned_move_id', 'in', moves.ids)])
for returned_moved in returned_moves:
    returned_moves_per_move[returned_moved.origin_returned_move_id] |= returned_moved

for order in self:
    pickings = self.env['stock.picking']
    for line order.order_line:
        moves = moves_per_line[line]
        for move in moves.ids:
            moves |= returned_moves_per_move[move]
        pickings |= moves.mapped('picking_id')
    order.picking_ids = pickings
    order.picking_count = len(pickings)

```