@Whenrow 2020-01-06
> @Whenrow
> The read_group method does not allow for grouping on 'id' (ambiguous query), and of course, without the possibility to group_by on it, the whole point of using the read_group is moot here.

Indeed you could replace this one by a search_read on stock.move to retrieve only ids