@mao-odoo 2020-01-16
It looks like my estimates of the gain of performance were completely off (new correct values have been updated in the commit message and the PR description).

With my example use case, the whole _compute_picking execution is :
 - with the current code :
      - 6.1 to 6.2 sec per pass
 - with @sle-odoo 's version (I have not tested it on runbot yet): 
      - 1.04 to 1.12 sec per pass
      -  --> 5.4 to 6x improvement 
 - with my SQL version : 
      - 0.02 to 0.04 sec per pass
      - -->   > 100x improvement

Of course, the difference on small datasets is negligible.

I feel like the performance gain is significant enough that it is worth it. What do you think about it @Whenrow ?