@sle-odoo 2020-01-17
It seems we all missed the override of the return wizard [here](8248f0e153ae5fcc1449d0d05149b61df615ad4f), meaning we could get rid of the second lookup on sock move to get the returned moves. So we could do this:

``` python
    @api.depends('order_line.move_ids.state', 'order_line.move_ids.picking_id')
    def _compute_picking(self):
        self.order_line.move_ids.flush(['state', 'picking_id'])
        query = """SELECT po.id, array_agg(sp.id)
                     FROM stock_move AS sm
               INNER JOIN purchase_order_line AS pol ON pol.id = sm.purchase_line_id
               INNER JOIN purchase_order AS po ON pol.order_id = po.id
               INNER JOIN stock_picking AS sp ON sp.id = sm.picking_id
                 GROUP BY po.id"""
        self.env.cr.execute(query)
        from collections import defaultdict
        pickings_per_po = defaultdict(list)
        for po_id, picking_ids in self._cr.fetchall():
            pickings_per_po[po_id] += picking_ids
        for order in self:
            order.picking_ids = self.env['stock.picking'].browse(pickings_per_po[order.id])
            order.picking_count = len(pickings_per_po[order.id])
```
The tradeoff is that it'll fetch the  returns of returns (and so on), which wasn't the case before the patch, you'll need to check with the product owners but for me it's acceptable