@mao-odoo 2020-01-20
@sle-odoo 
This is for a v12, so flush is not available and, afaik, there isn't really a suitable alternative. (I can only think of a env.cr.commit, but if there is an error afterward, we're ducked for the auto roll back)

If I understand correctly, your proposition is based on the fact that we have all the picking related information that we need just from the moves and thus do not need to take care of the returned_move_ids.

So I guess something like this would be good right?

```python
    @api.depends('order_line.move_ids.returned_move_ids',
                 'order_line.move_ids.state',
                 'order_line.move_ids.picking_id')
    def _compute_picking(self):
        for order in self:
            # We keep a limited scope on purpose. Ideally, we should also use move_orig_ids and
            # do some recursive search, but that could be prohibitive if not done correctly.
            moves = self.env['stock.move'].search_read(
                [('purchase_line_id', 'in', order.order_line.ids)],
                ['picking_id']
            )
            picking_ids = [move['picking_id'][0] for move in moves]
            pickings = self.env['stock.picking'].browse(picking_ids)
            order.picking_ids = pickings
            order.picking_count = len(pickings)
```