@sle-odoo 2020-01-20
yeah but after reading correctly everything i changed my mind

two weeks ago i said

> @mao-odoo are you sure the issue stems from the use of mapped and the prefetching ? are both mapped used in this methods slow? if it's indeed the case and due to the prefetch, did you try disabling the prefetch at this call?

because you said in your original message

> When doing a return on a picking with multiple hundreds of products,
the return processing would timeout.

now when i make a flamegraph on the last v12 i see that the time spent in mapped in less than 10%. I also missed the fact that this was a stored m2m, now that i know that (the information is computed ~slowly once then stored and becomes ~free) i'm really against adding more complexity there. Plus i think adding an index on the proper field (origin_returned_move_id?) would make these mapped faster and would be a better fix for a stable vesion.

why not tackling the 60% of the time in spent in actually creating the returned moves? in this part of the code, it seems [this](https://github.com/odoo/odoo/blob/12.0/addons/stock/wizard/stock_picking_return.py#L55) is pretty slow (around 20%, so more than the mapped you were looking at) and could be replaced by a read_group:sum.