@mao-odoo 2020-01-20
@sle-odoo 
Regarding the index:
You are absolutely right, I should have put that in this proposed fix too (that is actually the first thing I did on the DB where the perf issue appeared, it gave a 5 to 8 x perf gain)

Regarding your flamegraph:
It seems like we have widely different results on this.
Here is my flamegraph (on my specific use case) with the index on origin_returned_move_id already manually created. https://drive.google.com/file/d/1nBjBkVSR50H5qEI-ZzKpufZZv8OpIipZ/view  (this was generated last Thursday with the latest version of the code at the time)
As you can see, 75% of the time is spent in mapped.