@mao-odoo 2020-01-24
@sle-odoo 

Here is WAY simpler version based on your remarque about https://github.com/odoo/odoo/commit/8248f0e153ae5fcc1449d0d05149b61df615ad4f 

Performance figures:
tested with 663 products in the Reverse Transfer
- base code without index on origin_returned_move_id:
    - 26 sec per pass

- base code with an index:
    - 5.4 to 6.2 sec

- with https://github.com/odoo/odoo/pull/42528/commits/48d6afeecf24998fdc92e57526cad91b8ca983c8 
    - 0.12 to 0.21 sec